add prechecks for file-methods

+++ src/files/file_methods.cpp
@@ -37,7 +37,7 @@ bool
 isFile(const std::string filePath)
 {
     if(boost::filesystem::exists(filePath) == false
-            || boost::filesystem::is_directory(filePath))
+            || boost::filesystem::is_regular(filePath) == false)
     {
         return false;
     }
@@ -178,6 +178,12 @@ renameFileOrDir(const std::string &oldPath,
                 const std::string &newPath,
                 std::string &errorMessage)
 {
+    if(doesPathExist(oldPath) == false)
+    {
+        errorMessage = "source-path " + oldPath + " doesn't exist.";
+        return false;
+    }
+
     boost::system::error_code boostError;
     boost::filesystem::rename(oldPath, newPath, boostError);
 
@@ -207,6 +213,12 @@ copyPath(const std::string &sourcePath,
          std::string &errorMessage,
          const bool force)
 {
+    if(doesPathExist(sourcePath) == false)
+    {
+        errorMessage = "source-path " + sourcePath + " doesn't exist.";
+        return false;
+    }
+
     boost::system::error_code boostError;
     if(force) {
         boost::filesystem::remove_all(targetPath);
@@ -234,6 +246,12 @@ bool
 createDirectory(const std::string &path,
                 std::string &errorMessage)
 {
+    if(isFile(path))
+    {
+        errorMessage = "under path " + path + " a file already exist.";
+        return false;
+    }
+
     boost::system::error_code boostError;
     const bool result = boost::filesystem::create_directories(path, boostError);
 
@@ -250,12 +268,16 @@ createDirectory(const std::string &path,
  * @param path path to delete
  * @param errorMessage reference for error-message output
  *
- * @return true, if successful, else false
+ * @return true, if successful, else false. Also return true, if path is already deleted.
  */
 bool
 deleteFileOrDir(const std::string &path,
                 std::string &errorMessage)
 {
+    if(doesPathExist(path) == false) {
+        return true;
+    }
+
     boost::system::error_code boostError;
     const bool result = boost::filesystem::remove_all(path, boostError);
